+++ lib/appfuel/application/dispatcher.rb
@@ -7,6 +7,8 @@ module Appfuel
           container[:feature_initializer].call(request.feature, container)
           action   = container[:action_loader].call(request.namespace, container)
           response = action.run(request.inputs)
+        rescue Appfuel::Handler::HandlerFailure => e
+          response = e.response
         rescue => e
           handler = Appfuel::ResponseHandler.new
           response = handler.error(e)