+++ lib/appfuel/handler.rb
@@ -1,6 +1,7 @@
 require_relative 'handler/validator_dsl'
 require_relative 'handler/inject_dsl'
 
+require_relative 'handler/handler_failure'
 require_relative 'handler/base'
 require_relative 'handler/action'
 require_relative 'handler/command'