+++ lib/appfuel/initializers/all.rb
@@ -1,4 +1,4 @@
 require_relative 'logging'
 require_relative 'db'
 require_relative 'web_api'
-require_relative 'aws_dynamodb'
+require_relative 'dynamodb'