+++ lib/appfuel/storage/aws_dynamodb.rb
@@ -1,2 +1,0 @@
-require_relative 'aws_dynamodb/repository'
-require_relative 'aws_dynamodb/no_sql_model'