+++ lib/appfuel/storage/dynamodb.rb
@@ -1,0 +1,2 @@
+require_relative 'dynamodb/repository'
+require_relative 'dynamodb/adapter'